some simplifications in the backend
