+++ DeepINN/__init__.py
@@ -4,7 +4,7 @@ from .__about__ import __version__
 from . import backend
 # from . import callbacks
 # from . import data
-# from . import geometry
+from . import geometry
 # from . import gradients as grad
 # from . import BC
 # from . import nn
