--- DeepINN/utils/array_ops_compat.py
@@ -1,1 +1,0 @@
-pass
