--- DeepINN/utils/external.py
@@ -1,1 +1,0 @@
-pass
