--- DeepINN/utils/internal.py
@@ -1,1 +1,0 @@
-pass
