--- DeepINN/utils/pytorch.py
@@ -1,1 +1,0 @@
-pass
