+++ test.py
@@ -2,6 +2,8 @@ import DeepINN as dp
 #print(dp)
 #print(dp.__version__)
 
-import DeepINN.backend as bkd
 
-#print(bkd.lib.__dict__.keys())
+print('dp keys ',dp.__dict__.keys())
+print(dp.__dict__["__about__"])
+
+#print()
